refactor: やっぱ model に移植（vmにする理由がない）
