+++ lib/model/data/participant.dart
@@ -1,3 +1,4 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'generated/participant.freezed.dart';
@@ -14,3 +15,20 @@ class Participant with _$Participant {
   factory Participant.fromJson(Map<String, dynamic> json) =>
       _$ParticipantFromJson(json);
 }
+
+class ParticipantsNotifier extends StateNotifier<List<Participant>> {
+  ParticipantsNotifier() : super([]);
+
+  void addParticipant(Participant participant) {
+    state = [...state, participant];
+  }
+
+  void removeParticipant(String id) {
+    state = state.where((user) => user.user_id != id).toList();
+  }
+}
+
+final participantsProvider =
+    StateNotifierProvider<ParticipantsNotifier, List<Participant>>((ref) {
+  return ParticipantsNotifier();
+});
